refactor: move dao package to database
