+++ app/src/androidTest/java/org/feature/fox/coffee_counter/data/local/dao/ItemDaoTest.kt
@@ -9,6 +9,7 @@ import com.google.common.truth.Truth.assertThat
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.test.runTest
 import org.feature.fox.coffee_counter.data.local.ItemDatabase
+import org.feature.fox.coffee_counter.data.local.database.dao.ItemDao
 import org.feature.fox.coffee_counter.data.local.database.tables.Item
 import org.feature.fox.coffee_counter.getOrAwaitValue
 import org.junit.After
