+++ app/src/main/java/org/feature/fox/coffee_counter/data/repository/UserRepository.kt
@@ -1,7 +1,7 @@
 package org.feature.fox.coffee_counter.data.repository
 
 import androidx.lifecycle.LiveData
-import org.feature.fox.coffee_counter.data.local.dao.UserDao
+import org.feature.fox.coffee_counter.data.local.database.dao.UserDao
 import org.feature.fox.coffee_counter.data.local.database.tables.Funding
 import org.feature.fox.coffee_counter.data.local.database.tables.Purchase
 import org.feature.fox.coffee_counter.data.local.database.tables.User
